add AddAssign, move new_gas_cost_summary_from_txn_effects()
dprint fmt
