+++ crates/iota-core/src/checkpoints/mod.rs
@@ -22,6 +22,7 @@ use iota_macros::fail_point;
 use iota_metrics::{MonitoredFutureExt, monitored_future, monitored_scope};
 use iota_network::default_iota_network_config;
 use iota_protocol_config::ProtocolVersion;
+use iota_sdk_types::GasCostSummary;
 use iota_types::{
     base_types::{AuthorityName, ConciseableName, EpochId, TransactionDigest},
     committee::StakeUnit,
@@ -30,7 +31,6 @@ use iota_types::{
     effects::{TransactionEffects, TransactionEffectsAPI},
     error::{IotaError, IotaResult},
     event::SystemEpochInfoEvent,
-    gas::{GasCostSummary, new_gas_cost_summary_from_txn_effects},
     iota_system_state::{
         IotaSystemState, IotaSystemStateTrait,
         epoch_start_iota_system_state::EpochStartSystemStateTrait,
@@ -1627,7 +1627,7 @@ impl CheckpointBuilder {
             .map(|c| (c.epoch, c.epoch_rolling_gas_cost_summary.clone()))
             .unwrap_or_default();
         let current_gas_costs =
-            new_gas_cost_summary_from_txn_effects(cur_checkpoint_effects.iter());
+            checked::new_gas_cost_summary_from_txn_effects(cur_checkpoint_effects.iter());
         if previous_epoch == self.epoch_store.epoch() {
             // sum only when we are within the same epoch
             GasCostSummary::new(
@@ -2521,6 +2521,43 @@ impl CheckpointServiceNotify for CheckpointService {
     }
 }
 
+#[iota_macros::with_checked_arithmetic]
+mod checked {
+    use iota_sdk_types::GasCostSummary;
+    use iota_types::effects::{TransactionEffects, TransactionEffectsAPI};
+    use itertools::MultiUnzip;
+
+    #[expect(clippy::type_complexity)]
+    pub fn new_gas_cost_summary_from_txn_effects<'a>(
+        transactions: impl Iterator<Item = &'a TransactionEffects>,
+    ) -> GasCostSummary {
+        let (
+            storage_costs,
+            computation_costs,
+            computation_costs_burned,
+            storage_rebates,
+            non_refundable_storage_fee,
+        ): (Vec<u64>, Vec<u64>, Vec<u64>, Vec<u64>, Vec<u64>) = transactions
+            .map(|e| {
+                (
+                    e.gas_cost_summary().storage_cost,
+                    e.gas_cost_summary().computation_cost,
+                    e.gas_cost_summary().computation_cost_burned,
+                    e.gas_cost_summary().storage_rebate,
+                    e.gas_cost_summary().non_refundable_storage_fee,
+                )
+            })
+            .multiunzip();
+
+        GasCostSummary::new(
+            computation_costs.iter().sum(),
+            computation_costs_burned.iter().sum(),
+            storage_costs.iter().sum(),
+            storage_rebates.iter().sum(),
+            non_refundable_storage_fee.iter().sum(),
+        )
+    }
+}
 // test helper
 pub struct CheckpointServiceNoop {}
 impl CheckpointServiceNotify for CheckpointServiceNoop {
