+++ crates/iota-types/src/mock_checkpoint_builder.rs
@@ -11,7 +11,7 @@ use crate::{
     committee::Committee,
     crypto::{AuthoritySignInfo, AuthoritySignature, IotaAuthoritySignature},
     effects::{TransactionEffects, TransactionEffectsAPI},
-    gas::{GasCostSummary, add_gas_cost_summary},
+    gas::GasCostSummary,
     messages_checkpoint::{
         CertifiedCheckpointSummary, CheckpointContents, CheckpointSummary,
         CheckpointVersionSpecificData, EndOfEpochData, FullCheckpointContents, VerifiedCheckpoint,
@@ -62,10 +62,7 @@ impl MockCheckpointBuilder {
         transaction: VerifiedTransaction,
         effects: TransactionEffects,
     ) {
-        add_gas_cost_summary(
-            &mut self.epoch_rolling_gas_cost_summary,
-            effects.gas_cost_summary(),
-        );
+        self.epoch_rolling_gas_cost_summary += effects.gas_cost_summary();
 
         self.transactions
             .push(VerifiedExecutionData::new(transaction, effects))
